Transfered TransformCoordinates class to Utils namespace
